# mock: add Counters, Gauges and Timers methods to get a map of all stats